+++ src/cells/SetCell.h
@@ -53,8 +53,6 @@ public:
   std::unique_ptr<Cell> Copy(GroupCell *group) const override;
   const CellTypeInfo &GetInfo() override;
 
-  using Cell::SetCurrentPoint;
-  void SetCurrentPoint(wxPoint point) const override;
   void Draw(wxDC *dc, wxDC *antialiassingDC) override;
 
   wxString ToMatlab() const override;